# [Fabric] Optimize cloning ShadowTree when animating layout props